separate lib and component from plugin and view
