+++ extension/cockpit-plugin/src/main/frontend/lib/camunda-plugin.d.ts
