+++ extension/cockpit-plugin/src/main/frontend/lib/configuration.ts
