+++ extension/cockpit-plugin/src/main/frontend/lib/message.ts
